pyproject.toml
[build-system]

requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aiida-fourstate"
version = "0.1.0"
description = "AiiDA common workflow to compute magnetic coupling constants with the four-state method"
authors = [{ name = "Giovanni Pizzi" }]
dependencies = ["aiida-core>=2.7", "numpy", "ase", "aiida_common_workflows"]

[project.entry-points.'aiida.workflows']
"fourstate.fourstate" = "aiida_fourstate:MagneticExchangeWorkChain"
